# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Voucher API Documentation",
        "description": "API documentation for Voucher Management System",
        "contact": {
            "email": "admin@example.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Local API Server"
        }
    ],
    "paths": {
...
        "/api/rewards/{user_id}": {
            "get": {
                "tags": [
                    "Rewards"
                ],
                "summary": "Get user's rewards history",
                "description": "Get history of completed leg rewards for a user. A reward is created when all three legs reach $1500.",
                "operationId": "getUserRewards",
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "path",
                        "description": "Unique user ID (e.g. MC34234)",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "start_date",
                        "in": "query",
                        "description": "Start date for filtering (Y-m-d)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "end_date",
                        "in": "query",
                        "description": "End date for filtering (Y-m-d)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Rewards history retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "user": {
                                            "properties": {
                                                "user_id": {
                                                    "type": "string",
                                                    "example": "MC34234"
                                                },
                                                "registration_date": {
                                                    "type": "string",
                                                    "format": "date",
                                                    "example": "2024-01-01"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "rewards": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "date": {
                                                        "type": "string",
                                                        "format": "date",
                                                        "example": "2024-02-19"
                                                    },
                                                    "leg_a_balance": {
                                                        "description": "Total balance of leg A when reward was earned",
                                                        "type": "number",
                                                        "format": "float",
                                                        "example": 1500
                                                    },
                                                    "leg_b_balance": {
                                                        "description": "Total balance of leg B when reward was earned",
                                                        "type": "number",
                                                        "format": "float",
                                                        "example": 1500
                                                    },
                                                    "leg_c_balance": {
                                                        "description": "Total balance of leg C when reward was earned",
                                                        "type": "number",
                                                        "format": "float",
                                                        "example": 1500
                                                    },
                                                    "reward_amount": {
                                                        "description": "Reward amount received",
                                                        "type": "number",
                                                        "format": "float",
                                                        "example": 500
                                                    },
                                                    "is_rewarded": {
                                                        "description": "Whether the reward was paid",
                                                        "type": "boolean",
                                                        "example": true
                                                    },
                                                    "completed_at": {
                                                        "description": "When the reward was completed",
                                                        "type": "string",
                                                        "format": "date-time",
                                                        "example": "2024-02-19T00:00:00Z"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "User not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/rewards/{user_id}/summary": {
            "get": {
                "tags": [
                    "Rewards"
                ],
                "summary": "Get user's rewards summary",
                "description": "Get summary of user's rewards including total rewards and current leg balances",
                "operationId": "getUserRewardsSummary",
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "path",
                        "description": "Unique user ID (e.g. MC34234)",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Rewards summary retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "total_rewards": {
                                            "description": "Total rewards received",
                                            "type": "number",
                                            "format": "float",
                                            "example": 1500
                                        },
                                        "current_leg_balances": {
                                            "properties": {
                                                "leg_a": {
                                                    "description": "Current balance of leg A",
                                                    "type": "number",
                                                    "format": "float",
                                                    "example": 1500
                                                },
                                                "leg_b": {
                                                    "description": "Current balance of leg B",
                                                    "type": "number",
                                                    "format": "float",
                                                    "example": 1500
                                                },
                                                "leg_c": {
                                                    "description": "Current balance of leg C",
                                                    "type": "number",
                                                    "format": "float",
                                                    "example": 1500
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "last_reward_date": {
                                            "description": "Date of last reward received",
                                            "type": "string",
                                            "format": "date",
                                            "example": "2024-02-19"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "User not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
...
}